COMMERCE-686 Move classes to a better package name
